Main/GuestCtrl: Move the guest additions update code from RTIsoFs* to the appropriate RTVfs* APIs (not tested because of bugref:8078) [build fix]
svn:sync-xref-src-repo-rev: r123512

+++ src/VBox/Main/src-client/GuestSessionImplTasks.cpp
@@ -1,4 +1,4 @@
-/* $Id: GuestSessionImplTasks.cpp 72958 2018-07-07 21:27:53Z alexander.eichner@oracle.com $ */
+/* $Id: GuestSessionImplTasks.cpp 72960 2018-07-07 21:31:08Z alexander.eichner@oracle.com $ */
 /** @file
  * VirtualBox Main - Guest session tasks.
  */
@@ -1900,7 +1900,7 @@ int GuestSessionTaskUpdateAdditions::copyFileToGuest(GuestSession *pSession, RTV
     int rc = RTVfsFileOpen(hVfsIso, strFileSource.c_str(), RTFILE_O_OPEN | RTFILE_O_READ, &hVfsFile);
     if (RT_SUCCESS(rc))
     {
-        size_t cbSrcSize = 0;
+        uint64_t cbSrcSize = 0;
         rc = RTVfsFileGetSize(hVfsFile, &cbSrcSize);
         if (RT_SUCCESS(rc))
         {
